documentation added to main

+++ main.py
@@ -2,7 +2,17 @@ import shutil
 import os
 import routes
 
-def calcular_numero_archivos(directorio):
+
+def calcular_numero_archivos(directorio:str):
+    """
+    Calcula el número de archivos en un directorio dado.
+
+    Parámetros:
+    - directorio (str): La ruta al directorio del cual se desea calcular el número de archivos.
+
+    Retorna:
+    - int: El número de archivos en el directorio. Retorna -1 si el directorio no existe.
+    """
     if not os.path.exists(directorio):
         print(f"El directorio {directorio} no existe.")
         return -1
@@ -16,7 +26,16 @@ def calcular_numero_archivos(directorio):
 
     return contador_archivos
 
-def calcular_logs_en_origen(directorio):
+def calcular_logs_en_origen(directorio:str):
+    """
+    Calcula el número de archivos .log en un directorio dado.
+
+    Parámetros:
+    - directorio (str): La ruta al directorio del cual se desea calcular el número de archivos .log.
+
+    Retorna:
+    - int: El número de archivos en el directorio. Retorna -1 si el directorio no existe.
+    """
     if not os.path.exists(directorio):
         print("origen de datos incorrecto")
         return -1
@@ -29,7 +48,15 @@ def calcular_logs_en_origen(directorio):
             contador_logs += 1
     return contador_logs
 
-def mover_archivos(origen, destino):
+def mover_archivos(origen:str, destino:str, maximo_de_archivos:int):
+    """
+    Mueve los archivos .log de un directorio a otro siempre que haya espacio suficiente
+
+    Parámetros:
+    - Directorios de origen i destino
+    - Maximo de archivos .log permitidos en carpeta de destino
+
+    """
     if not os.path.exists(destino):
         os.makedirs(destino)
 
@@ -37,7 +64,7 @@ def mover_archivos(origen, destino):
     num_logs_origen = calcular_logs_en_origen(origen)
    
 
-    while num_archivos_destino < 250 and num_logs_origen>0:
+    while num_archivos_destino < maximo_de_archivos and num_logs_origen > 0:
 
         archivos = os.listdir(origen)
 
@@ -45,12 +72,12 @@ def mover_archivos(origen, destino):
             if archivo.endswith('.log'):
                 origen_archivo = os.path.join(origen, archivo)
                 destino_archivo = os.path.join(destino, archivo)
-                #if os.path.isfile(origen_archivo):  # Solo mover si es un archivo
-                if not os.path.exists(destino_archivo):  # Verificar si el archivo ya existe en el destino
-                    shutil.move(origen_archivo, destino_archivo)
-                    print(f"Archivo {archivo} movido a {destino}")
-                    num_archivos_destino += 1
-                    num_logs_origen = calcular_logs_en_origen(origen)
+                if os.path.isfile(origen_archivo):  # Solo mover si es un archivo (avitamos alguna carpeta que acabe como .log)
+                    if not os.path.exists(destino_archivo):  # Verificar si el archivo ya existe en el destino
+                        shutil.move(origen_archivo, destino_archivo)
+                        print(f"Archivo {archivo} movido a {destino}")
+                        num_archivos_destino += 1
+                        num_logs_origen = calcular_logs_en_origen(origen)
             if num_archivos_destino >= 250:
                print("carpeta de destino llena")
                break
@@ -62,4 +89,4 @@ carpeta_origen = routes.carpeta_origen
 carpeta_destino = routes.carpeta_destino
 
 # Llama a la función para mover archivos
-mover_archivos(carpeta_origen, carpeta_destino)
+mover_archivos(carpeta_origen, carpeta_destino, 250)
